+++ server/Cargo.toml
@@ -13,7 +13,7 @@ documentation = "https://docs.rs/jsonrpsee-ws-server"
 futures-channel = "0.3.14"
 futures-util = { version = "0.3.14", default-features = false, features = ["io", "async-await-macro"] }
 jsonrpsee-types = { path = "../types", version = "0.15.1" }
-jsonrpsee-core = { path = "../core", version = "0.15.1", features = ["server", "soketto"] }
+jsonrpsee-core = { path = "../core", version = "0.15.1", features = ["server", "soketto", "http-helpers"] }
 tracing = "0.1.34"
 serde = "1"
 serde_json = { version = "1", features = ["raw_value"] }
@@ -23,7 +23,7 @@ tokio-util = { version = "0.7", features = ["compat"] }
 tracing-futures = "0.2.5"
 tokio-stream = "0.1.7"
 http = "0.2.7"
-hyper = { version = "0.14", features = ["server"] }
+hyper = { version = "0.14", features = ["server", "http1", "http2"] }
 tower = "0.4.13"
 
 [dev-dependencies]
@@ -32,3 +32,4 @@ jsonrpsee-test-utils = { path = "../test-utils" }
 jsonrpsee = { path = "../jsonrpsee", features = ["full"] }
 rand = "0.8"
 tracing-subscriber = { version = "0.3.3", features = ["env-filter"] }
+tower = { version = "0.4.13", features = ["timeout"] }
