+++ server/src/tests/ws.rs
@@ -50,12 +50,12 @@ async fn can_set_the_max_request_body_size() {
 	let mut client = WebSocketTestClient::new(addr).await.unwrap();
 
 	// Invalid: too long
-	let req = format!(r#"{{"jsonrpc":"2.0", "method":{}, "id":1}}"#, "a".repeat(100));
+	let req = format!(r#"{{"jsonrpc":"2.0","method":"{}","id":1}}"#, "a".repeat(100));
 	let response = client.send_request_text(req).await.unwrap();
 	assert_eq!(response, oversized_request(100));
 
 	// Max request body size should not override the max response body size
-	let req = r#"{"jsonrpc":"2.0", "method":"anything", "id":1}"#;
+	let req = r#"{"jsonrpc":"2.0","method":"anything","id":1}"#;
 	let response = client.send_request_text(req).await.unwrap();
 	assert_eq!(response, ok_response(JsonValue::String("a".repeat(100)), Id::Num(1)));
 
@@ -132,6 +132,9 @@ async fn can_set_max_connections() {
 
 	// Decrement connection count
 	drop(conn2);
+
+	tokio::time::sleep(std::time::Duration::from_millis(100)).await;
+
 	// Can connect again
 	let conn4 = WebSocketTestClient::new(addr).await;
 	assert!(conn4.is_ok());
